+++ src/lib/ai.ts
@@ -1,12 +1,10 @@
 // cspell:words lstripBlocks
-import fs from 'fs'
 import OpenAI from 'openai'
 import { ChatCompletionCreateParamsBase } from 'openai/resources/chat/completions'
-import { join } from 'path'
 import nunjucks from 'nunjucks'
 import { ZodSchema } from 'zod'
 import { zodToJsonSchema } from 'zod-to-json-schema'
-import { Logger } from './interfaces'
+import { Logger, PromptFile } from './interfaces'
 import { makeAction, SendAction } from './bosun/action'
 import { sleep } from 'openai/core'
 
@@ -48,7 +46,7 @@ export interface LLMStep {
      * Prompt can be a simple string or a link to a file, loaded with `loadFile` function which
      * takes a path to the file relative to `src/use-cases` directory. Should be Nunjucks-compatible.
      */
-    prompt: string | File
+    prompt: string | PromptFile
 
     /**
      * Schema for structured LLM output using {@link zod https://zod.dev/}
@@ -212,13 +210,6 @@ export interface AIEngine {
     createStep: <T extends LLMStep | ProgrammaticStep>(step: T) => T
 
     /**
-     * Loads a file from the specified path.
-     * @param path - The path to the file to load.
-     * @returns The loaded File object.
-     */
-    loadFile: (path: string) => File
-
-    /**
      * Creates a new conversation instance.
      * @param messages - Optional initial messages for the conversation.
      * @returns A new Conversation object.
@@ -334,10 +325,6 @@ export interface Message {
     text: string
     /** Optional URL of an image associated with the message */
     imageUrl?: string
-}
-
-export interface File {
-    content: () => Promise<string>
 }
 
 /**
@@ -632,21 +619,9 @@ export function createAIEngine(cfg: EngineConfig = {}): AIEngine {
         return response.choices[0].message.content
     }
 
-    function loadFile(path: string) {
-        // NOTE: there probably will be S3 loading stuff here
-
-        return {
-            content: async () => {
-                logger.debug('AI Engine: loading prompt:', path)
-                return fs.promises.readFile(join(basePath, path), 'utf-8')
-            },
-        }
-    }
-
     return {
         createWorkflow: createWorkflow,
         createStep,
-        loadFile,
         createConversation: getConversation,
     }
 }